mandi-mcp/requirements.txt
mcp
httpx
pydantic
python-dotenv
google-generativeai
gTTS
fastapi
uvicorn
python-multipart